microservices/auth-service/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
PyJWT==2.8.0